src/data_science_on_databricks/transformations/sample_trips_my_project.py
from databricks.sdk.runtime import spark
from pyspark import pipelines as dp
from pyspark.sql import DataFrame

# This file defines a sample transformation.
# Edit the sample below or add new transformations
# using "+ Add" in the file browser.


@dp.table
def sample_trips_my_project() -> DataFrame:
    return spark.read.table("samples.nyctaxi.trips")
