pyproject.toml
[project]
    name = "data-science-on-databricks"
    version = "0.1.0"
    description = "Data Science on Databricks"
    authors = [
        {name = "J. van Nijnatten"},
    ]
    readme = "README.md"
    license = "MIT"
    requires-python = ">=3.14"
    classifiers = [
        "License :: Osi Approved :: MIT License",
        "Operating System :: OS Independent",
        "Private :: Do Not Upload",
        "Programming Language :: Python :: 3.14",
    ]
    dependencies = [
        "databricks-connect>=18.0.0",
        "databricks-dlt>=0.3.0",
        "pandas>=2.3.3",
        "pyyaml>=6.0.3",
        "rich>=14.2.0",
        "textual>=6.6.0",
        "typer>=0.20.0",
]

[dependency-groups]
    dev = [
        "pre-commit==4.5.0",
        {include-group = "lint"},
        {include-group = "test"},
    ]
    lint = [
        "ruff==0.14.6",
        "mypy==1.18.2",
        "bandit==1.9.1",
        "radon>=6.0.1",
]
    test = [
        "pytest==9.0.1",
        "pytest-cov>=7.0.0",
        "pytest-mock==3.15.1",
]

[build-system]
    requires = ["uv_build>=0.9.11,<0.10.0"]
    build-backend = "uv_build"

[project.scripts]
    main = "data_science_on_databricks.main:main"

[tool.uv]
    package = true

[tool.ruff]
    lint.select = [
        "A",   # flake8-builtins
        "ANN", # flake8-annotations
        "B",   # flake8-bugbear
        "BLE", # flake8-blind-except
        "C",   # flake8-comprehensions
        "COM", # flake8-commas
        "C90", # mccabe
        "E",   # pycodestyle errors
        "F",   # pyflakes
        "G",   # flake8-logging-format
        "I",   # isort
        "PD",  # pandas-vet
        "RUF", # ruff-specific rules
        "S5",  # flake8-bandit
        "T10", # flake&-debugger
        "T20", # flake8-print
        "W",   # pycodestyle warnings
    ]

    lint.ignore = [
        "W191",   # tab-indentation,
        "E111",   # indentation-with-invalid-multiple
        "E114",   # indentation-with-invalid-multiple-comment
        "E117",   # over-indented
        "D206",   # indent-with-spaces
        "D300",   # triple-single-quotes
        "Q000",   # bad-quotes-inline-string
        "Q001",   # bad-quotes-multiline-string
        "Q002",   # bad-quotes-docstring
        "Q003",   # avoidable-escaped-quote
        "COM812", # missing-trailing-comma
        "COM819", # prohibited-trailing-comma
        "ISC001", # single-line-implicit-string-concatenation
        "ISC002", # multi-line-implicit-string-concatenation
        "ANN204", # missing-return-type-special-method
        "S105",   # hardcoded-password-string
        "E501",   # line-too-long
    ]

    exclude = [
        "tests/data"
    ]


[tool.ruff.format]
    quote-style = "double"
    indent-style = "space"
    line-ending = "auto"
    docstring-code-format = true
    docstring-code-line-length = 50


[tool.ruff.lint.per-file-ignores]
    "tests/**/*.py" = [
        "S101", # flake8-bandit: assert
    ]
    "__init__.py" = [
        "F401",
    ]
    "notebook_run unit_tests.py" = [
        "F821",
        "E402",
        "S101"
    ]


[tool.mypy]
    exclude = [
        "tests",
    ]
    ignore_missing_imports = true


[tool.bandit]
    exclude_dirs = [
        "tests"
    ]


[tool.radon]
    order = "SCORE"
    cc_min = "C"


[tool.pytest]
    addopts = [
        "--strict-markers"
    ]
    markers = [
        "slow",
        "databricks",
    ]
    testpaths = [
        "tests"
    ]
